+++ .github/workflows/test.yaml
@@ -24,3 +24,4 @@ jobs:
         run: |
           npm run build
           npm run lint
+          npm run test
