+++ packages/query-composer/tsconfig.json
@@ -8,9 +8,8 @@
     "noEmit": false,
     "outDir": "dist/types",
     "resolveJsonModule": true,
-    "rootDir": "src",
     "skipLibCheck": true,
     "strict": true
   },
-  "include": ["src"]
+  "include": ["src", "tests"]
 }
